requirements.txt
fastapi
uvicorn[standard]
python-dotenv

openai
tiktoken

langchain
langchain-community
chromadb

pypdf
python-multipart
httpx
numpy

psycopg2-binary
sqlalchemy
pgvector

openpyxl
pandas
